feat: progress on https://github.com/tracking-exposed/YCAI/issues/66

+++ backend/bin/videofetch.js
@@ -7,34 +7,63 @@ const path = require('path');
 
 const CSV = require('../lib/CSV');
 const curly = require('../lib/curly');
+const DEFAULT_EXPERIMENT = "experiments";
 
 nconf.argv().env().file({ file: 'config/settings.json'});
 
+function guardoniFormat(e, i) {
+  // the current entry is a videoId and a title
+  return {
+    url: `https://www.youtube.com/watch?v=${e.videoId}`,
+    watchFor: "30s",
+    urltag: `${i}΅ ${e.title}`,
+  }
+}
+
+function simpleFormat(e) {
+  return {
+    videoURL: `https://www.youtube.com/watch?v=${e.videoId}`,
+    title: e.title,
+  }
+}
+
 async function vfet(channelId) {
-  const { html, statusCode }= await curly.fetchRawChannelVideosHTML(channelId);
+  let selectedcontent, htmlfile, dfile = null;
+
+  if(!!nconf.get('guardoni'))
+    dfile = path.join(DEFAULT_EXPERIMENT, `guardoni-${channelId}.csv`);
+  else
+    dfile = path.join(DEFAULT_EXPERIMENT, `${channelId}-urls.csv`);
+
+  if(fs.existsSync(dfile))
+    return console.log(`File ${dfile} exists, quitting!`);
+
+  htmlfile = path.join(DEFAULT_EXPERIMENT, `${channelId}-dump.html`);
+  if(fs.existsSync(htmlfile))
+    return console.log(`HTML dump file ${htmlfile} exists, quitting!`);
+
+  const { html, statusCode } = await curly.fetchRawChannelVideosHTML(channelId);
 
   debug("Status %d", statusCode);
-  const titlesandId = await curly.recentVideoFetch(channelId)
+  const titlesandId = await curly.parseRecentVideosHTML(html);
 
   if(!titlesandId) {
     debug("Failure in extracting video details from channel %s", channelId);
-    return { json: { error: true, message: "Failure in extracting info from YouTube; investigate"}}
+    return;
   }
 
-  const urltitle = _.map(titlesandId, function(e) {
-    return {
-      videoURL: 'https://www.youtube.com/watch?v=' + e.videoId,
-      title: e.title,
-    }
-  });
+  if(!!nconf.get('guardoni'))
+    selectedcontent = _.map(titlesandId, guardoniFormat);
+  else
+    selectedcontent = _.map(titlesandId, simpleFormat);
 
-  const csvcontent = CSV.produceCSVv1(urltitle);
-  const dfile = path.join('experiments', channelId + '-urls.csv');
+  const csvcontent = CSV.produceCSVv1(selectedcontent);
   fs.writeFileSync(dfile, csvcontent);
   debug("Produced %s with %d URLs and title",
-    dfile, urltitle.length);
+    dfile, selectedcontent.length);
 };
 
 if(!nconf.get('channel'))
-  return console.log("Mandatory --channel <channelId>");
+  return console.log("Mandatory --channel <channelId>, <--guardoni> is optional");
+
 vfet(nconf.get('channel'));
